refactor: create common topic dirs
can't tell how this'll be in the long run, when many files will have
story and test files. but for now i guess it seems like the files are a
bit more balanced across dirs
