+++ web/src/modules/topic/components/CriteriaTable/NodeCell.styles.tsx
@@ -1,6 +1,6 @@
 import styled from "@emotion/styled";
 
-import { nodeWidth } from "../EditableNode/EditableNode.styles";
+import { nodeWidth } from "../Node/EditableNode.styles";
 
 export const StyledDiv = styled.div`
   height: 100%; // expand to fill cell in case other cells are bigger due to nodes having more rows of text
