+++ web/src/modules/topic/components/CriteriaTable/NodeCell.tsx
@@ -1,5 +1,5 @@
 import { Node } from "../../utils/diagram";
-import { EditableNode } from "../EditableNode/EditableNode";
+import { EditableNode } from "../Node/EditableNode";
 import { StyledDiv } from "./NodeCell.styles";
 
 export const NodeCell = ({ node }: { node: Node }) => {
