+++ web/src/modules/topic/components/Indicator/CriteriaIndicator.tsx
