+++ web/src/modules/topic/components/Indicator/CriteriaTableIndicator.tsx
