+++ web/src/modules/topic/components/Indicator/EdgeClaimIndicator.tsx
