+++ web/src/modules/topic/components/Indicator/NodeClaimIndicator.tsx
